move current main page to explorer page
